First commit: Version command & onready event.
